Removed eval function
