+++ src/main/scala/shine/DPIA/Phrases/Phrase.scala
@@ -2,15 +2,13 @@ package shine.DPIA.Phrases
 
 import arithexpr.arithmetic.{NamedVar, RangeAdd}
 import shine.DPIA.Lifting.{liftDependentFunction, liftFunction, liftPair}
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics.{IndexData, NatData}
 import shine.DPIA.Types._
 import shine.DPIA.Types.TypeCheck._
 import shine.DPIA._
 import shine.DPIA.primitives.functional.NatAsIndex
 
 sealed trait Phrase[T <: PhraseType] {
-  val t: T // TODO? perform type checking at the same time
+  val t: T
 }
 
 final case class Identifier[T <: PhraseType](name: String, `type`: T)
@@ -117,7 +115,7 @@ final case class BinOp(op: Operators.Binary.Value, lhs: Phrase[ExpType], rhs: Ph
     }
 }
 
-final case class Literal(d: OperationalSemantics.Data)
+final case class Literal(d: Data)
   extends Phrase[ExpType] {
 
   assert(!d.isInstanceOf[NatData])
@@ -370,17 +368,12 @@ sealed trait Primitive[T <: PhraseType] extends Phrase[T] {
     throw new Exception("visitAndRebuild should be implemented by a macro")
 }
 
-trait ExpPrimitive extends Primitive[ExpType] {
-  def eval(s: OperationalSemantics.Store): OperationalSemantics.Data = ???
-}
+trait ExpPrimitive extends Primitive[ExpType]
 
-trait AccPrimitive extends Primitive[AccType] {
-  def eval(s: OperationalSemantics.Store): OperationalSemantics.AccIdentifier = ???
-}
+trait AccPrimitive extends Primitive[AccType]
 
 trait CommandPrimitive extends Primitive[CommType] {
   override val t: CommType = comm
-  def eval(s: OperationalSemantics.Store): OperationalSemantics.Store = ???
 }
 
 object Operators {
