+++ src/main/scala/shine/DPIA/primitives/functional/Idx.scala
@@ -1,8 +1,6 @@
 package shine.DPIA.primitives.functional
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types.DataType._
 import shine.DPIA.Types._
 import shine.DPIA._
@@ -17,12 +15,4 @@ final case class Idx(n: Nat,
   index :: expT(idx(n), read)
   array :: expT(n`.`dt, read)
   override val t: ExpType = expT(dt, read)
-
-  override def eval(s: Store): Data = {
-    (OperationalSemantics.eval(s, array),
-     OperationalSemantics.eval(s, index)) match {
-      case (ArrayData(xs), IntData(i)) => xs(i)
-      case _ => throw new Exception("This should not happen")
-    }
-  }
 }
