+++ src/main/scala/shine/DPIA/primitives/functional/Join.scala
@@ -1,8 +1,6 @@
 package shine.DPIA.primitives.functional
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types.DataType._
 import shine.DPIA.Types._
 import shine.DPIA._
@@ -17,17 +15,4 @@ final case class Join(n: Nat,
                      ) extends ExpPrimitive {
   array :: expT(n`.`(m`.`dt), w)
   override val t: ExpType = expT((n * m)`.`dt, w)
-
-  override def eval(s: Store): Data = {
-    OperationalSemantics.eval(s, array) match {
-      case ArrayData(outer) =>
-        val arrays = outer.map {
-          case ArrayData(inner) => inner
-          case _ => throw new Exception("This should not happen")
-        }
-        ArrayData(arrays.flatten)
-
-      case _ => throw new Exception("This should not happen")
-    }
-  }
 }
