+++ src/main/scala/shine/DPIA/primitives/functional/MakePair.scala
@@ -1,8 +1,6 @@
 package shine.DPIA.primitives.functional
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types._
 import shine.DPIA._
 import shine.macros.Primitive.expPrimitive
@@ -17,10 +15,4 @@ final case class MakePair(dt1: DataType,
   fst :: expT(dt1, access)
   snd :: expT(dt2, access)
   override val t: ExpType = expT(dt1 x dt2, access)
-
-  override def eval(s: Store): Data = {
-    PairData(
-      OperationalSemantics.eval(s, fst),
-      OperationalSemantics.eval(s, snd))
-  }
 }
