+++ src/main/scala/shine/DPIA/primitives/functional/NatAsIndex.scala
@@ -1,8 +1,6 @@
 package shine.DPIA.primitives.functional
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics.{IndexData, NatData}
 import shine.DPIA.Types.DataType._
 import shine.DPIA.Types._
 import shine.DPIA._
@@ -14,11 +12,4 @@ final case class NatAsIndex(n: Nat,
                            ) extends ExpPrimitive {
   e :: expT(NatType, read)
   override val t: ExpType = expT(idx(n), read)
-
-  override def eval(s: OperationalSemantics.Store): OperationalSemantics.Data = {
-    OperationalSemantics.eval(s, e) match {
-      case NatData(m) => IndexData(m, n)
-      case d => throw new Exception(s"Expected NatData but found $d.")
-    }
-  }
 }
