+++ src/main/scala/shine/DPIA/primitives/imperative/DropAcc.scala
@@ -1,13 +1,10 @@
 package shine.DPIA.primitives.imperative
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types._
 import shine.DPIA.Types.DataType._
 import shine.DPIA._
 import shine.macros.Primitive.accPrimitive
-
-import scala.xml.Elem
 
 // this drops n many elements from an array of m elements
 @accPrimitive
