+++ src/main/scala/shine/DPIA/primitives/imperative/IdxVecAcc.scala
@@ -1,8 +1,6 @@
 package shine.DPIA.primitives.imperative
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types.DataType._
 import shine.DPIA.Types._
 import shine.DPIA._
@@ -17,13 +15,4 @@ final case class IdxVecAcc(n: Nat,
   index :: expT(idx(n), read)
   vector :: accT(vec(n, st))
   override val t: AccType = accT(st)
-
-  override def eval(s: Store): AccIdentifier = {
-    val vectorE = OperationalSemantics.eval(s, vector)
-    val indexE = OperationalSemantics.eval(s, index) match {
-      case IntData(i) => i
-      case _ => throw new Exception("This should not happen")
-    }
-    VectorAccessIdentifier(vectorE, indexE)
-  }
 }
