+++ src/main/scala/shine/DPIA/primitives/imperative/PairAcc2.scala
@@ -1,8 +1,6 @@
 package shine.DPIA.primitives.imperative
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types._
 import shine.DPIA._
 import shine.macros.Primitive.accPrimitive
@@ -14,11 +12,4 @@ final case class PairAcc2(dt1: DataType,
                          ) extends AccPrimitive {
   pair :: accT(dt1 x dt2)
   override val t: AccType = accT(dt2)
-
-  override def eval(s: Store): AccIdentifier = {
-    OperationalSemantics.eval(s, pair) match {
-      case r: PairIdentifier => r.snd
-      case _ => throw new Exception("This should not happen")
-    }
-  }
 }
