+++ src/main/scala/shine/DPIA/primitives/imperative/Seq.scala
@@ -1,8 +1,6 @@
 package shine.DPIA.primitives.imperative
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types._
 import shine.macros.Primitive.comPrimitive
 
@@ -12,9 +10,4 @@ final case class Seq(c1: Phrase[CommType],
   extends CommandPrimitive {
   c1 :: comm
   c2 :: comm
-
-  override def eval(s: Store): Store = {
-    val s1 = OperationalSemantics.eval(s, c1)
-    OperationalSemantics.eval(s1, c2)
-  }
 }
