+++ src/main/scala/shine/DPIA/primitives/imperative/Skip.scala
@@ -1,7 +1,6 @@
 package shine.DPIA.primitives.imperative
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types._
 
 import scala.xml.Elem
@@ -10,8 +9,6 @@ import scala.xml.Elem
 case class Skip() extends CommandPrimitive {
 
   override val t: CommType = comm
-
-  override def eval(s: Store): Store = s
 
   override def visitAndRebuild(f: VisitAndRebuild.Visitor): Phrase[CommType] = this
