+++ src/main/scala/shine/OpenCL/Compilation/Passes/FlagPrivateArrayLoops.scala
@@ -1,11 +1,10 @@
 package shine.OpenCL.Compilation.Passes
 
-import shine.DPIA.Nat
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics.ArrayData
 import shine.DPIA.Types.{CommType, PhraseType}
 import shine.DPIA.primitives.functional.{Idx, NatAsIndex}
 import shine.DPIA.primitives.imperative.{For, ForNat, IdxAcc}
+import shine.DPIA.{ArrayData, Nat}
 import shine.OpenCL.AddressSpace
 import shine.OpenCL.primitives.imperative.{New, ParFor, ParForNat}
