+++ src/main/scala/shine/OpenCL/primitives/functional/CircularBuffer.scala
@@ -1,11 +1,9 @@
 package shine.OpenCL.primitives.functional
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types.DataType._
 import shine.DPIA.Types._
 import shine.DPIA._
-import shine.DPIA.primitives.functional.Slide
 import shine.macros.Primitive.expPrimitive
 
 @expPrimitive
@@ -21,9 +19,4 @@ final case class CircularBuffer(a: AddressSpace,
   load :: expT(dt1, read) ->: expT(dt2, write)
   input :: expT((n - 1 + sz)`.`dt1, read)
   override val t: ExpType = expT(n`.`(sz`.`dt2), read)
-
-  override def eval(s: Store): Data = {
-    import shine.DPIA.primitives.functional.Map
-    Slide(n, sz, 1, dt2, Map((n - 1 + sz), dt1, dt2, read, load, input)).eval(s)
-  }
 }
