+++ src/main/scala/shine/OpenCL/primitives/functional/RotateValues.scala
@@ -1,7 +1,6 @@
 package shine.OpenCL.primitives.functional
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types.DataType._
 import shine.DPIA.Types._
 import shine.DPIA._
@@ -18,9 +17,4 @@ final case class RotateValues(a: AddressSpace,
   write :: expT(dt, read) ->: expT(dt, shine.DPIA.Types.write)
   input :: expT((n - 1 + sz)`.`dt, read)
   override val t: ExpType = expT(n`.`(sz`.`dt), read)
-
-  override def eval(s: Store): Data = {
-    import shine.DPIA.primitives.functional.Slide
-    Slide(n, sz, 1, dt, input).eval(s)
-  }
 }
