+++ src/main/scala/shine/OpenCL/primitives/functional/ToMem.scala
@@ -1,8 +1,6 @@
 package shine.OpenCL.primitives.functional
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types._
 import shine.DPIA._
 import shine.macros.Primitive.expPrimitive
@@ -14,6 +12,4 @@ final case class ToMem(addrSpace: AddressSpace,
                       ) extends ExpPrimitive {
   input :: expT(dt, write)
   override val t: ExpType = expT(dt, read)
-
-  override def eval(s: Store): Data = OperationalSemantics.eval(s, input)
 }
