+++ src/main/scala/shine/OpenCL/primitives/imperative/New.scala
@@ -1,9 +1,6 @@
 package shine.OpenCL.primitives.imperative
 
-import shine.DPIA.DSL.identifier
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types._
 import shine.DPIA._
 import shine.macros.Primitive.comPrimitive
@@ -14,11 +11,4 @@ final case class New(a: AddressSpace,
                      f: Phrase[VarType ->: CommType]
                     ) extends CommandPrimitive {
   f :: varT(dt) ->: comm
-
-  override def eval(s: Store): Store = {
-    val f_ = OperationalSemantics.eval(s, f)
-    val arg = identifier(freshName("x"), f.t.inT)
-    val newStore = OperationalSemantics.eval(s + (arg.name -> 0), f_(arg))
-    newStore - arg.name
-  }
 }
