+++ src/main/scala/shine/OpenMP/primitives/functional/MapPar.scala
@@ -1,8 +1,6 @@
 package shine.OpenMP.primitives.functional
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types.DataType._
 import shine.DPIA.Types._
 import shine.DPIA._
@@ -18,16 +16,4 @@ final case class MapPar(n: Nat,
   f :: expT(dt1, read) ->: expT(dt2, write)
   array :: expT(n`.`dt1, read)
   override val t: ExpType = expT(n`.`dt2, write)
-
-  override def eval(s: Store): Data = {
-    val fE = OperationalSemantics.eval(s, f)
-    OperationalSemantics.eval(s, array) match {
-      case ArrayData(xs) =>
-        ArrayData(xs.map { x =>
-          OperationalSemantics.eval(s, fE(Literal(x)))
-        })
-
-      case _ => throw new Exception("This should not happen")
-    }
-  }
 }
