+++ src/main/scala/shine/OpenMP/primitives/functional/ReducePar.scala
@@ -1,8 +1,6 @@
 package shine.OpenMP.primitives.functional
 
 import shine.DPIA.Phrases._
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types.DataType._
 import shine.DPIA.Types._
 import shine.DPIA._
@@ -19,17 +17,4 @@ final case class ReducePar(n: Nat,
   init :: expT(dt2, write)
   array :: expT(n`.`dt1, read)
   override val t: ExpType = expT(dt2, read)
-
-  override def eval(s: Store): Data = {
-    val fE = OperationalSemantics.eval(s, f)(BinaryFunctionEvaluator)
-    val initE = OperationalSemantics.eval(s, init)
-    OperationalSemantics.eval(s, array) match {
-      case ArrayData(xs) =>
-        ArrayData(Vector(xs.fold(initE) {
-          (x, y) => OperationalSemantics.eval(s,
-            fE(Literal(x))(Literal(y)))
-        }))
-      case _ => throw new Exception("This should not happen")
-    }
-  }
 }
