+++ src/main/scala/shine/OpenMP/primitives/imperative/ParFor.scala
@@ -1,13 +1,9 @@
 package shine.OpenMP.primitives.imperative
 
-import shine.DPIA.DSL._
 import shine.DPIA.Phrases.{CommandPrimitive, Phrase, _}
-import shine.DPIA.Semantics.OperationalSemantics
-import shine.DPIA.Semantics.OperationalSemantics._
 import shine.DPIA.Types.DataType._
 import shine.DPIA.Types._
 import shine.DPIA._
-import shine.DPIA.primitives.functional.NatAsIndex
 import shine.macros.Primitive.comPrimitive
 
 @comPrimitive
@@ -22,15 +18,5 @@ final case class ParFor(n: Nat,
   lazy val unwrapBody: (Identifier[ExpType], Identifier[AccType], Phrase[CommType]) = body match {
     case Lambda(i, Lambda(o, body)) => (i, o, body)
     case _ => throw new Exception("This should not happen")
-  }
-
-  override def eval(s: Store): Store = {
-    val nE = evalIndexExp(s, NatAsIndex(n, Natural(n)))
-    val bodyE = OperationalSemantics.eval(s, body)(OperationalSemantics.BinaryFunctionEvaluator)
-
-    (0 until nE.eval).foldLeft(s)((s1, i) => {
-      OperationalSemantics.eval(s1,
-        bodyE(Literal(i))(out `@` Literal(i)))
-    })
   }
 }
